+++ packages/krb5/env/run.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+HOME=`pwd dirname "${BASH_SOURCE}"`
+
+cd $HOME/krb5
+docker-compose up --abort-on-container-exit