+++ packages/krb5/lib/delprinc.js
@@ -5,14 +5,14 @@
 
 // ## Options
 
+// * `admin.server`   
+//   Address of the kadmin server; optional, use "kadmin.local" if missing.   
+// * `admin.principal`   
+//   KAdmin principal name unless `kadmin.local` is used.   
+// * `admin.password`   
+//   Password associated to the KAdmin principal.   
 // * `principal`   
 //   Principal to be created.   
-// * `kadmin_server`   
-//   Address of the kadmin server; optional, use "kadmin.local" if missing.   
-// * `kadmin_principal`   
-//   KAdmin principal name unless `kadmin.local` is used.   
-// * `kadmin_password`   
-//   Password associated to the KAdmin principal.   
 // * `keytab`   
 //   Path to the file storing key entries.   
 
@@ -23,37 +23,60 @@
 // .krb5_delrinc({
 //   principal: 'myservice/my.fqdn@MY.REALM',
 //   keytab: '/etc/security/keytabs/my.service.keytab',
-//   kadmin_principal: 'me/admin@MY_REALM',
-//   kadmin_password: 'pass',
-//   kadmin_server: 'localhost'
+//   admin: {
+//     principal: 'me/admin@MY_REALM',
+//     password: 'pass',
+//     server: 'localhost'
+//   }
 // }, function(err, status){
-//   console.log(err ? err.message : 'Principal removed: ' + status);
+//   console.info(err ? err.message : 'Principal removed: ' + status);
 // });
 // ```
 
-// ## Source Code
-var misc;
+// ## Hooks
+var handler, misc, mutate, on_options;
 
-module.exports = function({options}) {
-  var cmd_delprinc, cmd_getprinc;
+on_options = function({options}) {
+  // Import all properties from `options.krb5`
+  if (options.krb5) {
+    mutate(options, options.krb5);
+    return delete options.krb5;
+  }
+};
+
+// ## Handler
+handler = function({options}) {
   if (!options.principal) {
     throw Error('Property principal is required');
   }
-  if (/.*@.*/.test(options.kadmin_principal)) {
+  if (/.*@.*/.test(options.admin.principal)) {
     // Normalize realm and principal for later usage of options
     if (options.realm == null) {
-      options.realm = options.kadmin_principal.split('@')[1];
+      options.realm = options.admin.principal.split('@')[1];
     }
   }
   if (!/^\S+@\S+$/.test(options.principal)) {
     options.principal = `${options.principal}@${options.realm}`;
   }
   // Prepare commands
-  cmd_getprinc = misc.kadmin(options, `getprinc ${options.principal}`);
-  cmd_delprinc = misc.kadmin(options, `delprinc -force ${options.principal}`);
-  this.system.execute({
-    cmd: cmd_delprinc,
-    if_exec: `${cmd_getprinc} | grep '${options.principal}'`
+  this.krb5.execute({
+    options: {
+      admin: options.admin,
+      cmd: `getprinc ${options.principal}`,
+      egrep: new RegExp(`^.*${misc.regexp.escape(options.principal)}$`)
+    },
+    metadata: {
+      shy: true
+    }
+  });
+  this.krb5.execute({
+    if: function() {
+      return this.status(-1);
+    },
+    options: {
+      admin: options.admin,
+      cmd: `delprinc -force ${options.principal}`
+    }
   });
   return this.system.remove({
     target: options.keytab,
@@ -61,5 +84,13 @@ module.exports = function({options}) {
   });
 };
 
+// ## Export
+module.exports = {
+  handler: handler,
+  on_options: on_options
+};
+
 // ## Dependencies
 misc = require('@nikitajs/core/lib/misc');
+
+({mutate} = require('mixme'));